+++ histoprint/__init__.py
@@ -4,4 +4,9 @@ from histoprint.formatter import HistFormatter, print_hist, text_hist
 
 from .version import version as __version__
 
-__all__ = ("print_hist", "text_hist", "formatter", "HistFormatter", "__version__")
+__all__ = (
+    "print_hist",
+    "text_hist",
+    "HistFormatter",
+    "__version__",
+)