+++ tests/test_basic.py
@@ -34,13 +34,13 @@ def test_hist():
         symbols="      ",
         bg_colors="rgbcmy",
         labels="ABCDE",
+        max_count=600.0,
     )
     hp.print_hist(
         histAll,
         title="Summaries",
         symbols=r"=|\/",
-        fg_colors="0",
-        bg_colors="0",
+        use_color=False,
         labels=["AAAAAAAAAAAAAAAA", "B", "CCCCCCCCCCCCC", "D"],
         summary=True,
     )
@@ -49,6 +49,33 @@ def test_hist():
         title="No composition",
         labels=["A", "B", "C"],
     )
+
+
+def test_width():
+    """Test output width."""
+    hist = np.histogram(np.random.randn(100) / 1000)
+    f = io.StringIO()
+    hp.print_hist(hist, file=f, columns=30, use_color=False)
+    f.flush
+    f.seek(0)
+    n_max = 0
+    for line in f:
+        print(line, end="")
+        n = len(line.rstrip())
+        n_max = np.max((n, n_max))
+    assert n_max == 30
+
+    hist = (np.array((100.5, 17.5)), np.array((0, 1, 2)))
+    f = io.StringIO()
+    hp.print_hist(hist, file=f, columns=30, use_color=False)
+    f.flush
+    f.seek(0)
+    n_max = 0
+    for line in f:
+        print(line, end="")
+        n = len(line.rstrip())
+        n_max = np.max((n, n_max))
+    assert n_max == 30
 
 
 def test_nan():
@@ -118,3 +145,39 @@ def test_stack():
     hp.print_hist(hist_stack, file=out2, title="Overlays", labels="ABCD")
 
     assert out1.getvalue() == out2.getvalue()
+
+
+def test_rich_histogram():
+    """Test updating the values of a histogram object."""
+    rich = pytest.importorskip("rich")
+
+    from histoprint.rich import RichHistogram
+
+    A = np.random.randn(1000) - 2
+    B = np.random.randn(1000) + 2
+
+    hA = np.histogram(A, bins=np.linspace(-5, 5, 11))
+    hB = np.histogram(B, bins=hA[1])
+
+    hist = RichHistogram(hA, columns=30)
+
+    rich.print(hist)
+
+    # Update values
+    count = hA[0]
+    count += hB[0]
+
+    rich.print(hist)
+
+    from rich.table import Table
+
+    tab = Table(title="Test table")
+    tab.add_column("left justify", justify="left", width=29)
+    tab.add_column("center justify", justify="center", width=35)
+    tab.add_column("right justify", justify="right", width=35)
+
+    from rich.align import Align
+
+    tab.add_row(hist, Align.center(hist), Align.right(hist))
+
+    rich.print(tab)