+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Addded
+- RichHistogram class for use with the `rich` package.
+- Argument to set maximum count in bins for plotting.
+
+### Changed
+- Changed tick formatting.
+
 ## [2.3.0]
 
 ### Changed